moved pngCleaner to root dir so it will be included when installed via npm
