+++ pgn-parser.js
@@ -69,5 +69,5 @@ const parse = function(input, options) {
 module.exports = {
   SyntaxError: parser.SyntaxError,
   parse:       parse,
-  clean:       require('./src/pngCleaner')
+  clean:       require('./pngCleaner')
 };
