+++ pngCleaner.js
